refactor(cli): split swizzle into api/swizzle leaf shape
Reorganize the flat swizzle command into the fractal leaf shape:

- api/swizzle/list/list.mjs  -> swizzle.list
- api/swizzle/copy/copy.mjs  -> swizzle.copy receipt (owns rewriteImports)
- api/swizzle/_adapter.mjs   -> shared @astryxdesign/core discovery + component
                               listing used by both leaves (resolveCore)
- api/swizzle/swizzle.mjs    -> dispatcher + barrel; keeps the same exports
                               (swizzle, rewriteImports) so api/index.mjs, the
                               CLI command, and the unit tests are unchanged

Pure reorganization — no behavior change. Verified byte-identical human and
--json output for `swizzle --list`, a real `swizzle Button` copy, and the
`swizzle NotARealComponent` error (stdout, stderr, and exit code all match).
The CLI command, api/index.mjs, shared lib/ modules, and the central
types/swizzle declarations are untouched.
